+++ src/App.js
@@ -3,7 +3,7 @@ import {BrowserRouter, Switch, Route} from 'react-router-dom'
 import Index from './pages/index/Index'
 import Login from './pages/login/Login'
 import Signup from './pages/signup/Signup'
-import Convites from './pages/index/convites/Convites'
+
 import Error from './pages/Error'
 import NavbarEmpresa from './pages/index/NavbarEmpresa'
 import NavbarJogador from './pages/index/NavbarJogador'
@@ -14,7 +14,7 @@ import Streams from './pages/index/company/streams/Streams'
 import Championship from './pages/index/company/championship/Championship'
 import Invite from './pages/index/player/invite/Invite'
 import Campeonato from './pages/index/campeonatos/Campeonato'
-import NovaPartida from './pages/index/company/novaPartida/NovaPartida'
+import NovoTime from './pages/index/company/novoTime/NovoTime'
 import { Hidden, Grid } from '@material-ui/core'
 
 class App extends Component {
@@ -24,8 +24,8 @@ class App extends Component {
 				<Route path="/(index)" exact component={Index} />
 				<Route path='/login' component={Login}/>
 				<Route path='/signup' component={Signup}/>
-				<Route path="/(player|administro|campeonatos|campeonato|partida|detalhes)" component={NavbarJogador}/>
-				<Route path="/empresa/(campeonatos|streams|times|campeonato|partida|detalhes)" component={NavbarEmpresa}/>
+				<Route path="/(player|administro|campeonatos|campeonato|partida|detalhes|novotime)" component={NavbarJogador}/>
+				<Route path="/empresa/(campeonatos|streams|times|campeonato|partida|detalhes|novotime)" component={NavbarEmpresa}/>
 				<Switch>
 					{/* Jogador */}
 					<Route path="/" exact component={Index} />
@@ -34,16 +34,16 @@ class App extends Component {
 					<Route path="/player" component={Invite} />
 					<Route path="/campeonato" component={Campeonato}/>
 					<Route path='/administro' component={Championship}/>
-				
+
 					{/* Empresa */}
-					<Route path='/convites' component={Convites}/>
+
 					<Route path='/empresa/partida' component={Partida}/>
 					<Route path='/empresa/detalhes' component={Detalhes}/>
 					<Route path="/empresa/streams" component={Streams}/>
 					<Route path="/empresa/campeonato" component={Campeonato}/>
 					<Route path="/empresa/campeonatos" component={Championship}/>
 					<Route path="/convite" component={Invite}/>
-					<Route path="/nova_partida" component={NovaPartida}/>
+					<Route path="/empresa/novotime"component={NovoTime}/>
 					<Route component={Error} />
 				</Switch>
 				<Hidden smUp><div style={{height:'56px',widht:'100%'}} /><Grid container direction="row" justify="center" alignItems="flex-end"><Route path="/(player|campeonato|detalhes|streams|convites|campeonatos|convite)" component={BottomNavigationComponent}/></Grid></Hidden>
